Use more sophisticated approach to calculating image size
Now, when we figure out the size of the target image after assembling the chroot environment, we win a few and we lose a few. Pieman became a bit flexible since it doesn't create an image with the size of 7 GB for all occasions, but the target image may turn out not large enough for the rootfs. The previous approach to calculating of the target image size figured out (via du) how much space we need to store the rootfs and created the partition with the size. However, the approach didn't take into account the file system metadata size, so in some cases there was no enough space to store the rootfs. The current logic fixes it.