Do not reload systemd manager confirmation at startup
The code was inherited from rpi23-gen-image. On the one hand, there is no need to reload systemd manager configuration (if so, let one of the scripts from files/firstboot do it explicitly) and the set of commands seem to be redundant. On the other hand, some of the supported distributions (Devuan, for example) don't have systemd, so it's wrong to call systemctl without any check.